refactor: read the shell PATH from one place, move PATH scopes to platform
Four places wrote or read "the PATH a new shell gets": the rc-file edits
for the user scope, the profile.d snippet for the machine scope, the
registry pathScope for both Windows scopes, and the interception read
side doctor used. Install and doctor could disagree on what PATH looks
like, the class of bug the #454 review found.

internal/platform now holds the one read: NewShellPath returns the PATH
a new shell gets, and ShellPath.LookPath resolves a command over it and
reports which source held the winning directory, machine, user or a
shell profile. shim.InspectInterception builds on it, so the install
warning and doctor read one value.

The write side moves to platform too. The registry scopes are
platform.UserPath and platform.MachinePath, with Prepend, Append, Remove,
Contains and Entries. The Windows PATHEXT lookup and the environment
broadcast move with them. The profile.d bodies become
platform.WriteSystemProfile, RemoveSystemProfile and
SystemProfileInstalled, which the machine install on Unix calls with its
own path. shim holds no registry code and no profile.d code.

The user-scope rc-file edits stay in shim, because they are shell rc
logic in internal/alias, not registry or profile.d code. The Windows
registry tests move to internal/platform. shim tests that drive an
install redirect the scopes through RedirectUserPathForTest and
RedirectMachinePathForTest.

No behaviour changes.
